+++ core_packages/ecsly/lib/src/systems/systems.dart
@@ -1,8 +1,8 @@
 export '../plugins/plugin.dart';
 export '../plugins/plugin_registry.dart';
-export 'certified_job_system.dart';
 export 'ecs_observers.dart';
 export 'isolate_config.dart';
+export 'parallel_systems.dart';
 export 'phase_systems.dart';
 export 'query_extensions.dart';
 export 'schedule.dart';